Consolidate subprocess execution and rename ExecArgs
- Merge subprocess.py into subprocess_exec.py (single module)
- Rename ExecArgs to SubprocessExecArgs for clarity
- Remove DirectExecArgs alias and run_direct_exec wrapper
- Remove examples from ExecInput.create docstring
- Add sh_test to verify subprocess_exec has no fastmcp dependency
- Update all callers to use SubprocessExecArgs directly
