+++ mcp_infra/exec/bwrap.py
@@ -13,7 +13,7 @@ from mcp_infra.enhanced.flat_mixin import FlatTool
 from mcp_infra.enhanced.server import EnhancedFastMCP
 from mcp_infra.exec.models import BaseExecResult, ExecOutcome, TimeoutMs, render_outcome_to_result
 from mcp_infra.exec.read_image import ReadImageInput, validate_and_encode_image
-from mcp_infra.exec.subprocess import run_proc
+from mcp_infra.exec.subprocess_exec import run_proc
 
 BWRAP = os.getenv("BWRAP", "bwrap")
 ALLOW_UNSHARE_NET = os.getenv("DUCK_UNSHARE_NET", "0") == "1"
