+++ mcp_infra/exec/subprocess_exec.py
@@ -5,17 +5,74 @@ Provides exec tool without fastmcp dependency. Uses types from models.py.
 
 from __future__ import annotations
 
+import asyncio
 from pathlib import Path
 from typing import Any
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from mcp_infra.exec.models import BaseExecResult, TimeoutMs, render_outcome_to_result
-from mcp_infra.exec.subprocess import run_proc
+from mcp_infra.exec.models import (
+    BaseExecResult,
+    ExecOutcome,
+    ExecOutput,
+    Exited,
+    Killed,
+    TimedOut,
+    TimeoutMs,
+    async_timer,
+    render_outcome_to_result,
+)
 
 
-class ExecArgs(BaseModel):
-    """Arguments for shell command execution."""
+async def run_proc(
+    argv: list[str], timeout_s: float, *, cwd: Path | None = None, stdin: bytes | str | None = None
+) -> ExecOutcome:
+    """Run a subprocess with timeout, returning structured outcome."""
+    if stdin is None:
+        stdin_bytes: bytes | None = None
+    elif isinstance(stdin, str):
+        stdin_bytes = stdin.encode("utf-8", errors="replace")
+    else:
+        stdin_bytes = stdin
+
+    async with async_timer() as get_duration_ms:
+        proc = await asyncio.create_subprocess_exec(
+            *argv,
+            stdin=asyncio.subprocess.PIPE,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
+            cwd=cwd,
+        )
+
+        try:
+            out, err = await asyncio.wait_for(proc.communicate(input=stdin_bytes), timeout=timeout_s)
+
+            stdout_bytes = out if out is not None else b""
+            stderr_bytes = err if err is not None else b""
+            exit_code = proc.returncode if proc.returncode is not None else 0
+            output = ExecOutput(stdout=stdout_bytes, stderr=stderr_bytes)
+
+            # Detect if process was killed by signal (negative exit code on Unix)
+            if exit_code < 0:
+                signal_num = -exit_code
+                return ExecOutcome(output=output, exit=Killed(signal=signal_num), duration_ms=get_duration_ms())
+
+            return ExecOutcome(output=output, exit=Exited(exit_code=exit_code), duration_ms=get_duration_ms())
+
+        except TimeoutError:
+            proc.kill()
+            try:
+                out, err = await asyncio.wait_for(proc.communicate(), timeout=5)
+            except (TimeoutError, ProcessLookupError):
+                out, err = b"", b""
+            stdout_bytes = out if out is not None else b""
+            stderr_bytes = err if err is not None else b""
+            output = ExecOutput(stdout=stdout_bytes, stderr=stderr_bytes)
+            return ExecOutcome(output=output, exit=TimedOut(), duration_ms=get_duration_ms())
+
+
+class SubprocessExecArgs(BaseModel):
+    """Arguments for subprocess execution."""
 
     cmd: list[str] = Field(
         description="Command array passed directly to subprocess (no shell). "
@@ -36,16 +93,12 @@ def get_exec_tool_schema() -> dict[str, Any]:
         "type": "function",
         "name": "exec",
         "description": "Execute a shell command. Use for file operations, running tests, etc.",
-        "parameters": ExecArgs.model_json_schema(),
+        "parameters": SubprocessExecArgs.model_json_schema(),
     }
 
 
-async def run_exec(args: ExecArgs, *, default_cwd: Path | None = None) -> BaseExecResult:
+async def run_exec(args: SubprocessExecArgs, *, default_cwd: Path | None = None) -> BaseExecResult:
     """Execute a command and return the result.
-
-    Args:
-        args: Exec arguments (command, timeout, etc.)
-        default_cwd: Fallback working directory if args.cwd is not specified
 
     Raises:
         ValueError: If cmd is empty or contains non-strings
