+++ mcp_infra/exec/test_direct.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from fastmcp.client import Client
 
-from mcp_infra.exec.direct import DirectExecArgs, DirectExecServer
+from mcp_infra.exec.direct import DirectExecServer
+from mcp_infra.exec.subprocess_exec import SubprocessExecArgs
 from mcp_infra.exec.models import Exited
 from mcp_infra.testing.exec_stubs import DirectExecServerStub
 
@@ -13,6 +14,6 @@ async def test_direct_exec_echo_inproc() -> None:
     server = DirectExecServer()
     async with Client(server) as session:
         stub = DirectExecServerStub.from_server(server, session)
-        res = await stub.exec(DirectExecArgs(cmd=["/bin/echo", "hello"], max_bytes=100000, timeout_ms=5000))
+        res = await stub.exec(SubprocessExecArgs(cmd=["/bin/echo", "hello"], max_bytes=100000, timeout_ms=5000))
         assert res.exit == Exited(exit_code=0)
         assert res.stdout == "hello\n"
