+++ mcp_infra/testing/exec_stubs.py
@@ -1,7 +1,7 @@
 """Typed stubs for exec MCP servers."""
 
 from mcp_infra.exec.bwrap import BwrapExecArgs
-from mcp_infra.exec.direct import DirectExecArgs
+from mcp_infra.exec.subprocess_exec import SubprocessExecArgs
 from mcp_infra.exec.models import BaseExecResult
 from mcp_infra.exec.seatbelt import SandboxExecArgs, SandboxExecResult
 from mcp_infra.stubs.server_stubs import ServerStub
@@ -10,7 +10,7 @@ from mcp_infra.stubs.server_stubs import ServerStub
 class DirectExecServerStub(ServerStub):
     """Typed stub for direct (unsandboxed) exec server operations."""
 
-    async def exec(self, input: DirectExecArgs) -> BaseExecResult:
+    async def exec(self, input: SubprocessExecArgs) -> BaseExecResult:
         raise NotImplementedError  # Auto-wired at runtime
